requirements.txt
flask
locust
elastic-apm
psutil
redis
python-dotenv
gunicorn
prometheus_client
